+++ adapters/opsgenie/schedule/schedule_internal_test.go
@@ -4,8 +4,11 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"log"
+	"os"
 	"testing"
 
+	"github.com/opsgenie/opsgenie-go-sdk-v2/client"
 	"github.com/opsgenie/opsgenie-go-sdk-v2/og"
 	"github.com/opsgenie/opsgenie-go-sdk-v2/schedule"
 	"github.com/stretchr/testify/assert"
@@ -334,6 +337,7 @@ func TestSchedule_Remove(t *testing.T) { //nolint:funlen
 	})
 }
 
+//nolint:funlen
 func TestInit(t *testing.T) {
 	t.Parallel()
 
@@ -376,5 +380,35 @@ func TestInit(t *testing.T) {
 			assert.ErrorIs(t, err, gosync.ErrMissingConfig)
 			assert.ErrorContains(t, err, ScheduleID)
 		})
+	})
+
+	t.Run("with logger", func(t *testing.T) {
+		t.Parallel()
+
+		logger := log.New(os.Stderr, "custom logger", log.LstdFlags)
+
+		adapter, err := Init(ctx, map[gosync.ConfigKey]string{
+			OpsgenieAPIKey: "test",
+			ScheduleID:     "schedule",
+		}, WithLogger(logger))
+
+		assert.NoError(t, err)
+		assert.Equal(t, logger, adapter.(*Schedule).Logger)
+	})
+
+	t.Run("with client", func(t *testing.T) {
+		t.Parallel()
+
+		scheduleClient, err := schedule.NewClient(&client.Config{
+			ApiKey: "test",
+		})
+		assert.NoError(t, err)
+
+		adapter, err := Init(ctx, map[gosync.ConfigKey]string{
+			ScheduleID: "schedule",
+		}, WithClient(scheduleClient))
+
+		assert.NoError(t, err)
+		assert.Equal(t, scheduleClient, adapter.(*Schedule).client)
 	})
 }